Updated with full gameplay-text/tooltip translations
Updated/completed the English translation of this file, which includes tooltips, skill names and descriptions, Admiral names, and several other gameplay elements, as well as the whole GFWL process. Everything is accurate to the context in the game, as best as I can figure out, having played the full story mode.

Still needs to be run in Japanese locale mode, and lots of things still not localizable with current tools, but the RTS gameplay is now fully understandable to an English speaker!